requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
